WIP: Move each calculation into its own form
I wanted an easier way to find, add, and remove calculations. Making
each calculation their own form seemed like the cleanest way of doing
this.

Previously, all calculations were inside a single form. This made it
easy to integrate with the URL query parameters (an HTML form's default
button action is a GET, which automatically modifies the parameters) but
made the internal state of the form messy and hard to work with.